Crate ic_context_logger

Modules

macros
These macros provide an interface to ContextLogger

Macros

crit
Log a crit-level log, with context fields if given

debug
Log a debug-level log, with context fields if given

error
Log an error-level log, with context fields if given

fatal
Log a crit-level log, with context fields if given, then panic!()

info
Log an info-level log, with context fields if given

info_sample
Sample the given log event

log
Log an entry at the given log level, with context fields if given

log_metadata
Return a LogMetadata

new_logger
Create a new logger with the supplied context

trace
Log a trace-level message, with context fields if given

update_context

warn
Log a warn-level log, with context fields if given

Structs

ContextLogger
A logger that holds context that can be updated and logged

LogMetadata
Metadata about the log entry

Traits

Logger
Logs data of type T