Struct ic_context_logger::ContextLogger

pub struct ContextLogger<C, L> where
    C: Clone + Default,
    L: Logger<C>,  {
    pub context: C,
    pub inner_logger: L,
}

A logger that holds context that can be updated and logged

Methods of this trait should not be used directly, instead use the macros found in macros.rs

Fields
context: Cinner_logger: L
Implementations

impl<C, L> ContextLogger<C, L> where
C: Clone + Default,
L: Logger<C>,

pub fn new(logger: L) -> Self

pub fn get_context(&self) -> C

pub fn with_new_context(&self, context: C) -> Self

pub fn log(&self, message: String, context: C, metadata: LogMetadata)

pub fn is_enabled_at(&self, level: Level, module_path: &'static str) -> bool

pub fn should_sample<T: Into<u32>>(&self, key: String, value: T) -> bool

pub fn is_tag_enabled(&self, tag: String) -> bool

pub fn is_n_seconds<T: Into<i32>>(
&self,
seconds: T,
metadata: LogMetadata
) -> bool