Trait ic_context_logger::Logger

pub trait Logger<T>: Clone {
    fn log(&self, message: String, data: T, metadata: LogMetadata);

    fn is_enabled_at(&self, level: Level, module_path: &'static str) -> bool;

    fn should_sample<V: Into<u32>>(&self, key: String, value: V) -> bool;

    fn is_tag_enabled(&self, tag: String) -> bool;

    fn is_n_seconds<V: Into<i32>>(
        &self, 
        seconds: V, 
        metadata: LogMetadata
    ) -> bool;
}

Logs data of type T

Required methods

fn log(&self, message: String, data: T, metadata: LogMetadata)

Log the given log event

fn is_enabled_at(&self, level: Level, module_path: &'static str) -> bool

Return true if events should be logged at the given level and module, false otherwise

fn should_sample<V: Into<u32>>(&self, key: String, value: V) -> bool

Return true if a sample! log with the given key and value should be logged, false otherwise

fn is_tag_enabled(&self, tag: String) -> bool

Return true if a log with the given tag should be logged, false otherwise

fn is_n_seconds<V: Into<i32>>(&self, seconds: V, metadata: LogMetadata) -> bool

Return true if this is the first log in n seconds, false otherwise

Loading content...